+++ examples/hosts/app-integration/external-data/src/inventoryList.ts
